# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airport Ground Operations Optimizer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FlightFlightType =
  (typeof FlightFlightType)[keyof typeof FlightFlightType];

export const FlightFlightType = {
  Domestic: "Domestic",
  International: "International",
  Cargo: "Cargo",
} as const;

export type FlightStand = (typeof FlightStand)[keyof typeof FlightStand];

export const FlightStand = {
  "nose-in": "nose-in",
  remote: "remote",
  "tow-in": "tow-in",
} as const;

export type FlightStatus = (typeof FlightStatus)[keyof typeof FlightStatus];

export const FlightStatus = {
  Scheduled: "Scheduled",
  "Taxi-In": "Taxi-In",
  "At-Gate": "At-Gate",
  Boarding: "Boarding",
  Pushback: "Pushback",
  Departed: "Departed",
  Delayed: "Delayed",
} as const;

export type FlightDelayRisk =
  (typeof FlightDelayRisk)[keyof typeof FlightDelayRisk];

export const FlightDelayRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type FlightWeather = (typeof FlightWeather)[keyof typeof FlightWeather];

export const FlightWeather = {
  clear: "clear",
  rain: "rain",
  fog: "fog",
  storm: "storm",
} as const;

export interface Flight {
  id: string;
  flightNumber: string;
  airline: string;
  airlineCode: string;
  origin: string;
  destination: string;
  flightType: FlightFlightType;
  gate: string;
  stand: FlightStand;
  status: FlightStatus;
  scheduledArrival: string;
  actualArrival?: string | null;
  scheduledDeparture: string;
  nextMilestone: string;
  nextMilestoneAt: string;
  predictedTurnaroundMin: number;
  baselineTurnaroundMin: number;
  predictedDelayMin: number;
  delayRisk: FlightDelayRisk;
  bottleneckResource?: string | null;
  bottleneckReason?: string | null;
  crewAssigned: number;
  baggageVolume: number;
  congestionLevel: number;
  weather: FlightWeather;
}

export type FlightDetailTimelineItem = {
  ts: string;
  event: string;
  description: string;
};

export type FlightDetailFeatureContributionsItem = {
  feature: string;
  impactMin: number;
};

export type FlightDetail = Flight & {
  timeline: FlightDetailTimelineItem[];
  featureContributions: FlightDetailFeatureContributionsItem[];
};

export type ResourceUnitStatus =
  (typeof ResourceUnitStatus)[keyof typeof ResourceUnitStatus];

export const ResourceUnitStatus = {
  idle: "idle",
  busy: "busy",
  maintenance: "maintenance",
} as const;

export interface ResourceUnit {
  id: string;
  label: string;
  status: ResourceUnitStatus;
  assignment?: string | null;
  utilisationPercent: number;
}

export type ResourceGroupType =
  (typeof ResourceGroupType)[keyof typeof ResourceGroupType];

export const ResourceGroupType = {
  gate: "gate",
  tug: "tug",
  stairs: "stairs",
  belt: "belt",
  crew: "crew",
  bridge: "bridge",
} as const;

export interface ResourceGroup {
  type: ResourceGroupType;
  label: string;
  totalUnits: number;
  busyUnits: number;
  utilisationPercent: number;
  units: ResourceUnit[];
}

export interface UtilizationPoint {
  ts: string;
  gate: number;
  tug: number;
  belt: number;
  crew: number;
  bridge: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Alert {
  id: string;
  severity: AlertSeverity;
  resource: string;
  flightId?: string | null;
  flightNumber?: string | null;
  message: string;
  suggestedAction: string;
  etaMinutes: number;
  createdAt: string;
}

export type AnalyticsSummaryTurnaroundTrendItem = {
  hour: string;
  actual: number;
  predicted: number;
};

export type AnalyticsSummaryBottleneckHeatmapItem = {
  hour: number;
  resource: string;
  count: number;
};

export type AnalyticsSummaryAirlinePerformanceItem = {
  airline: string;
  onTime: number;
  flights: number;
};

export interface AnalyticsSummary {
  onTimePerformance: number;
  avgTurnaroundMin: number;
  resourceEfficiency: number;
  delaysAvoided: number;
  totalFlightsToday: number;
  bottlenecksResolved: number;
  turnaroundTrend: AnalyticsSummaryTurnaroundTrendItem[];
  bottleneckHeatmap: AnalyticsSummaryBottleneckHeatmapItem[];
  airlinePerformance: AnalyticsSummaryAirlinePerformanceItem[];
}

export type SimulatorRequestWeather =
  (typeof SimulatorRequestWeather)[keyof typeof SimulatorRequestWeather];

export const SimulatorRequestWeather = {
  clear: "clear",
  rain: "rain",
  fog: "fog",
  storm: "storm",
} as const;

export interface SimulatorRequest {
  flightId: string;
  crewAssigned: number;
  gate: string;
  weather: SimulatorRequestWeather;
  baggageVolume: number;
  congestionLevel: number;
}

export type SimulatorResponseNewDelayRisk =
  (typeof SimulatorResponseNewDelayRisk)[keyof typeof SimulatorResponseNewDelayRisk];

export const SimulatorResponseNewDelayRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type SimulatorResponseContributionsItem = {
  feature: string;
  impactMin: number;
};

export interface SimulatorResponse {
  flightId: string;
  baselinePredictionMin: number;
  newPredictionMin: number;
  deltaMin: number;
  newDelayRisk: SimulatorResponseNewDelayRisk;
  newBottleneck?: string | null;
  contributions: SimulatorResponseContributionsItem[];
}

export type OptimizerResultReassignmentsItem = {
  flightId: string;
  flightNumber: string;
  fromGate: string;
  toGate: string;
  reason: string;
  expectedSavingMin: number;
};

export interface OptimizerResult {
  reassignments: OptimizerResultReassignmentsItem[];
  totalSavingMin: number;
}

export type ListFlights200 = {
  airport: string;
  serverTime: string;
  flights: Flight[];
};

export type ListResources200 = {
  groups: ResourceGroup[];
  history: UtilizationPoint[];
};

export type ListAlerts200 = {
  alerts: Alert[];
};

export type ResetSimulation200 = {
  ok: boolean;
};