# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Airport Ground Operations Optimizer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flights/{flightId}:
    get:
      operationId: getFlight
      tags: [flights]
      parameters:
        - in: path
          name: flightId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FlightDetail" }
...
  /simulator/predict:
    post:
      operationId: simulatePrediction
      tags: [simulator]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SimulatorRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulatorResponse" }
...
components:
  schemas:
...
    Flight:
      type: object
      properties:
        id: { type: string }
        flightNumber: { type: string }
        airline: { type: string }
        airlineCode: { type: string }
        origin: { type: string }
        destination: { type: string }
        flightType:
          type: string
          enum: [Domestic, International, Cargo]
        gate: { type: string }
        stand:
          type: string
          enum: [nose-in, remote, tow-in]
        status:
          type: string
          enum: [Scheduled, Taxi-In, At-Gate, Boarding, Pushback, Departed, Delayed]
        scheduledArrival: { type: string }
        actualArrival: { type: string, nullable: true }
        scheduledDeparture: { type: string }
        nextMilestone: { type: string }
        nextMilestoneAt: { type: string }
        predictedTurnaroundMin: { type: number }
        baselineTurnaroundMin: { type: number }
        predictedDelayMin: { type: number }
        delayRisk:
          type: string
          enum: [low, medium, high]
        bottleneckResource:
          type: string
          nullable: true
        bottleneckReason:
          type: string
          nullable: true
        crewAssigned: { type: integer }
        baggageVolume: { type: integer }
        congestionLevel: { type: integer }
        weather:
          type: string
          enum: [clear, rain, fog, storm]
      required:
        - id
        - flightNumber
        - airline
        - airlineCode
        - origin
        - destination
        - flightType
        - gate
        - stand
        - status
        - scheduledArrival
        - scheduledDeparture
        - nextMilestone
        - nextMilestoneAt
        - predictedTurnaroundMin
        - baselineTurnaroundMin
        - predictedDelayMin
        - delayRisk
        - crewAssigned
        - baggageVolume
        - congestionLevel
        - weather
    FlightDetail:
      allOf:
        - $ref: "#/components/schemas/Flight"
        - type: object
          properties:
            timeline:
              type: array
              items:
                type: object
                properties:
                  ts: { type: string }
                  event: { type: string }
                  description: { type: string }
                required: [ts, event, description]
            featureContributions:
              type: array
              items:
                type: object
                properties:
                  feature: { type: string }
                  impactMin: { type: number }
                required: [feature, impactMin]
          required: [timeline, featureContributions]
...
    SimulatorRequest:
      type: object
      properties:
        flightId: { type: string }
        crewAssigned: { type: integer }
        gate: { type: string }
        weather:
          type: string
          enum: [clear, rain, fog, storm]
        baggageVolume: { type: integer }
        congestionLevel: { type: integer }
      required: [flightId, crewAssigned, gate, weather, baggageVolume, congestionLevel]
    SimulatorResponse:
      type: object
      properties:
        flightId: { type: string }
        baselinePredictionMin: { type: number }
        newPredictionMin: { type: number }
        deltaMin: { type: number }
        newDelayRisk:
          type: string
          enum: [low, medium, high]
        newBottleneck: { type: string, nullable: true }
        contributions:
          type: array
          items:
            type: object
            properties:
              feature: { type: string }
              impactMin: { type: number }
            required: [feature, impactMin]
      required:
        - flightId
        - baselinePredictionMin
        - newPredictionMin
        - deltaMin
        - newDelayRisk
        - contributions
...