# lib/api-zod/src/generated/types/flightDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airport Ground Operations Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { Flight } from "./flight";
import type { FlightDetailFeatureContributionsItem } from "./flightDetailFeatureContributionsItem";
import type { FlightDetailTimelineItem } from "./flightDetailTimelineItem";

export type FlightDetail = Flight & {
  timeline: FlightDetailTimelineItem[];
  featureContributions: FlightDetailFeatureContributionsItem[];
};