# lib/api-zod/src/generated/types/simulatorRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Airport Ground Operations Optimizer API
 * OpenAPI spec version: 0.1.0
 */
import type { SimulatorRequestWeather } from "./simulatorRequestWeather";

export interface SimulatorRequest {
  flightId: string;
  crewAssigned: number;
  gate: string;
  weather: SimulatorRequestWeather;
  baggageVolume: number;
  congestionLevel: number;
}